iio: jesd204: axi_jesd204_[rx|tx]: Fix measured device/link clock
Need to cast to u64 before muliplication

Fixes: 74518b367c33 ("iio: jesd204: axi_jesd204_[rx|tx]: Fix measured device/link clock")

(cherry picked from commit 14a49f9204cf959c0295253a9f32b4a82143cfa6)

+++ drivers/iio/jesd204/axi_jesd204_rx.c
@@ -166,7 +166,7 @@ static ssize_t axi_jesd204_rx_status_read(struct device *dev,
 		ret += scnprintf(buf + ret, PAGE_SIZE - ret,
 			"Measured Link Clock: off\n");
 	} else {
-		clock_rate = DIV_ROUND_CLOSEST_ULL(DIV_ROUND_CLOSEST(jesd->axi_clk_freq,
+		clock_rate = DIV_ROUND_CLOSEST_ULL((u64)DIV_ROUND_CLOSEST(jesd->axi_clk_freq,
 			1000) * clock_ratio, 1ULL << 16);
 
 		ret += scnprintf(buf + ret, PAGE_SIZE - ret,
@@ -197,7 +197,7 @@ static ssize_t axi_jesd204_rx_status_read(struct device *dev,
 			ret += scnprintf(buf + ret, PAGE_SIZE - ret,
 				"Measured Device Clock: off\n");
 		} else {
-			clock_rate = DIV_ROUND_CLOSEST_ULL(DIV_ROUND_CLOSEST(jesd->axi_clk_freq,
+			clock_rate = DIV_ROUND_CLOSEST_ULL((u64)DIV_ROUND_CLOSEST(jesd->axi_clk_freq,
 				1000) * clock_ratio, 1ULL << 16);
 
 			ret += scnprintf(buf + ret, PAGE_SIZE - ret,
